CommandLine Refactor (cont'd)
Moved two additional files to Shared project
Added post-build steps to copy csv to project build target folder
