+++ Shared/Resources/colours.csv
@@ -1,4 +1,4 @@
-Name,R,G,B
+Name,R,G,B - Colors from BrickOwl.com
 Aqua,181,211,214
 Black,0,0,0
 Blue,0,87,166
